+++ vite.config.ts
@@ -7,5 +7,5 @@ export default defineConfig({
   plugins: [react(), tailwindcss()],
   optimizeDeps: {
     include: ["html-to-image", "zustand"],
-  },
+  }
 });